requirements.txt
streamlit
pydantic
sqlalchemy
pyodbc
python-dotenv
langchain
langchain-openai
langchain-community
pandas